# SDPA Optimizations: Reading cur pos + page table sharded