### Problem description
Previously cur pos tensor and page table is always read from DRAM. This PR intializes the cur pos tensor, page table as L1 sharded on sub core grid and all ops directly read from L1. For the page table we cast the page table to uint16 to save L1 space since indices won't go above16K

### What's changed
- Updated SDPA reader, Paged cache update op, plus one op to use globally allocated L1 address of cur pos and page table tensor
- Updated text demo and demo decode to create the cur pos tensor directly on device rather on host
- `copy_host_to_device` is updated so that it can handle cases for tensors already created on device
- page table is now casted to uint16 during decode in text demo

### Perf Improvement (looking at model perf kernel durations, old taken from latest superset)
old 8650 ns -> new SDPA-model-kernel-avg: 7815.972222222223 ns
old 4450 ns -> new  PagedUpdateCache-model-kernel-avg: 3669.2673611111113 ns
old 842 ns -> new PlusOne_0-model_tail-kernel-avg: 428.71875 ns

E2E Perf: 72.6 tok/s/u -> 73.15 tok/s/u

### Checklist
- [ ] [APC](https://github.com/tenstorrent/tt-metal/actions/runs/17030443623) (All passing, demo failing just because throughput is too high)
- [ ] 6u [TG Pipeline](https://github.com/tenstorrent/tt-metal/actions/runs/17030448926)